Lots of new code. Don't know what it does, but it's lots.

+++ source/app/controllers/Api/BaseController.php
@@ -8,7 +8,6 @@ use Input;
 class BaseController
 extends Controller
 {
-
     protected $embed = "";
 
     protected $direction = "desc";
@@ -24,7 +23,7 @@ extends Controller
         $this->embed     = explode(",", Input::get("embed", $this->embed));
         $this->direction = Input::get("direction", $this->direction);
         $this->filter    = explode(",", Input::get("filter", $this->filter));
-        $this->limit     = Input::get("limit", $this->limit);
+        $this->limit     = (int) Input::get("limit", $this->limit);
         $this->order     = Input::get("order", $this->order);
     }
 }
